pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sas2dbx"
version = "0.1.0"
description = "SAS Query to Databricks Migrator — AI-first transpilation tool"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["sas", "databricks", "pyspark", "migration", "transpiler"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
]
dependencies = [
    "pyyaml>=6.0",
    "typer[all]>=0.12",
    "rich>=13.0",
    "anthropic>=0.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]
web = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",
]
databricks = [
    "databricks-sdk>=0.30",
]
all = [
    "sas2dbx[web,databricks]",
]

[project.scripts]
sas2dbx = "sas2dbx.cli:app"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]

[tool.pytest.ini_options]
testpaths = ["tests"]
